Replace scraping a list of articles with fetching from API.
Using `newspaper-book` to split the articles into paper sections
and `newspaperPageNumber` to sort them.

Instead of scraping a web page to get a list of links, then
fetching each article individually, we can fetch all the
articles from the API in one go (or maybe two, if there are
more than 200).

We currently still scrape a page to work out which date is
the current day's paper.