Improve CSS for embeds
We were applying styles to `.element-embed` assuming it was a video,
giving it enough `padding-bottom` to display a 16:9 video.

But maybe this has changed since we implmented it. Looks like videos
are now in `.element-atom`. And the `.element-embed` styles were making
an embed like the email signup on https://www.theguardian.com/money/2018/mar/27/london-property-slump-most-widespread-since-financial-crisis
be overly-tall, and the form covered the preceding text.

So, move the `.element-embed` styles to `.element-atom` and see how
that goes.