Fix error saving file when article ID is too long
There was an article with an ID of 300 characters
(books_2021_may_07_the-best-recent-science-fiction-and-fantasy-reviews-roundupsorrowland-by-rivers-solomon-merky-1299-polly-ho-yens-dark-lullaby-titan-899-sarah-pinskers-we-are-satellites-head-of-zeus-1899-the-kingdoms-by-natasha-pulley-bloomsbury-1499-the-cottingley-cuckoo-by-aj-elwood-titan-899)
which caused an IOError when saving the file locally, because we use
that ID as the filename (plus ".html") and the maximum filename
length is 255 characters.

So I gave up on using nice readable filenames and now use a hash
of the ID as the filename.